third_party/blink/web_tests/fast/events/composition-event-source-device-event-sender-expected.txt

Test that composition events should have a sourceCapabilities and its firesTouchEvents should always be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

compositionstart
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
compositionupdate
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
compositionend
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
PASS successfullyParsed is true

TEST COMPLETE
